Initial driver
mostly copied code from the hostpath example.
Start re-writing the NodeServer and adding tests as I go.

Right now I'm thinking that the NodeServer won't have to worry about
initiating the streaming of remote volumes from other baggageclaim
nodes. That will happen in the ControllerServer.

Same will happen for deletion. The node will only delete volumes it made
during NodeStageVolume. Any volumes made as the result of streaming a
volume from another worker will not be deleted by the NodeServer. It
will be the responsibility of the ControllerServer to reach out to the
node later and delete the volume it streamed to a particular node.

Signed-off-by: Taylor Silva <tsilva@pivotal.io>